chore(cleanup): reorganized animation case ladder
This change reorders the onAnimationChanged function to be a bit more
readable, and make it clearer how the cases flow into one another with
the special case handling. Also fixed an issue where almost all 4 tick
weapons were being controlled by the "show zamorakian spear" toggle,
which looked to be a bug in how the cases were flowing into the zspear
check.